requirements.txt
amqp==2.5.1
beautifulsoup4==4.8.1
billiard==3.6.1.0
celery==4.3.0
cffi==1.13.2
Django==2.2.6
django-bootstrap4==1.0.1
django-ckeditor==5.8.0
django-js-asset==1.2.2
django-tempus-dominus==5.1.2.9
gevent==1.4.0
greenlet==0.4.15
importlib-metadata==0.23
kombu==4.6.5
lxml==4.4.2
more-itertools==7.2.0
Pillow==6.2.1
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
soupsieve==1.9.5
SQLAlchemy==1.3.10
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
